third_party/WebKit/LayoutTests/editing/execCommand/insert_list/insert_list_in_link.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => assert_selection(
  '<div contenteditable><a href="#">^foo|</a></div>',
  'InsertUnorderedList',
  '<div contenteditable><ul><li><a href="#">^foo|</a></li></ul></div>'),
  'InsertUnorderedList in link');
</script>
